restructure recipe plugin
