+++ modules/recipe/model/recipe.js
@@ -1,10 +1,19 @@
-/**
- * @module Recipes
- */
 'use strict';
+
+/**
+ * @module recipe/model
+ */
 
 let server;
 const first = 0;
+
+/**
+ * Route represents a view.
+ * @typedef {Object} Route
+ * @property {String} method - http verb.
+ * @property {String} path - path or pattern for Route.
+ * @property {Object} settings - all other attributes set on route.
+ */
 
 /**
  * Tests a route to see if it is the home page of a Recipe
@@ -18,29 +27,22 @@ function filterRecipes (route) {
 }
 
 /**
- * Route represents a view.
- * @typedef {Object} Route
- * @property {String} method - http verb.
- * @property {String} path - path or pattern for Route.
- * @property {Object} settings - all other attributes set on route.
+ * Stores a reference to the Hapi Server
+ * @function setServer
+ * @param {Object} serve - Hapi JS server object
+ * @returns {Undefined} undefined
  */
+function setServer (serve) {
+    server = serve;
+}
 
-module.exports = {
-    /**
-     * Stores a reference to the Hapi Server
-     * @function setServer
-     * @param {Object} serve - Hapi JS server object
-     * @returns {Undefined} undefined
-     */
-    setServer (serve) {
-        server = serve;
-    },
-    /**
-     * Finds all of the Recipe routes
-     * @function list
-     * @returns {Array} {Array} of {Route}
-     */
-    list () {
-        return server.table()[first].table.filter(filterRecipes);
-    }
-};
+/**
+ * Finds all of the Recipe routes
+ * @function list
+ * @returns {Array} {Array} of {Route}
+ */
+function list () {
+    return server.table()[first].table.filter(filterRecipes);
+}
+
+module.exports = {setServer, list};
